requirements.txt
gradio
openai
faiss-cpu
sentence-transformers
numpy
pytest
PyGithub
python-dotenv
requests